Replace everything is a string internally
Replace everything is a string internally to its native type.
int, float, bool and string are supported.

inc & dec are now requires that the types are the same.

set l 0; inc l 5.5;

is no longer correct, but has to be changed to

set l 0.0; inc l 5.5;

or:

set l 0; inc l [int 5.5];

with a rounding loss.

New commands are "int" and "float" for conversions.

Signed-off-by: Jonas Aaberg <cja@lithops.se>